feat(starterkit): let a ViewModel reset an animator trigger
A trigger that was set and never consumed stays armed. The animation then
fires the moment its state becomes reachable - often seconds later, in a
state nobody connected to it. Nothing in the package could undo that.

AnimatorResetTriggerMonoBinder and AnimatorResetTriggerBinder can, and they
cost almost nothing to add because the plumbing was extracted rather than
copied. AnimatorTriggerMonoBinder and AnimatorTriggerBinder own the
IRelayCommand/Action handover, the parameter check and the enabled-state
notification; each leaf supplies one line saying which operation it performs.

That extraction is one of the copy-paste sites the audit named, and it is
the one where a copy would have been most likely to drift: the two binders
would have differed only in a single method call.

The file holding AnimatorSetTriggerMonoBinder keeps its GUID and its name -
the base moved to a new file instead. A MonoBehaviour's script GUID is what
scenes and prefabs reference, and renaming the wrong file would have turned
every existing binder into a missing script.

Tests: two more in the animator fixture.

Verification: recompile clean, EditMode 794.
