+++ Aspid.MVVM/Packages/tech.aspid.mvvm/StarterKit/Unity/Runtime/Binders/Animators/Mono/AnimatorResetTriggerMonoBinder.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 050f3eee094042a89c04e19ad2befa0d
