+++ Aspid.MVVM/Packages/tech.aspid.mvvm/StarterKit/Unity/Runtime/Binders/Animators/Mono/AnimatorTriggerMonoBinder.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 87d67771ab364f73949453a70f1aede4
