+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- **Resetting an animator trigger.** A trigger that was set and never consumed stays armed, and the animation fires the moment its state becomes reachable — often seconds later, in a state nobody connected to it. Nothing in the package could undo that. `AnimatorResetTriggerMonoBinder` and `AnimatorResetTriggerBinder` do, and they cost almost nothing to add because the command plumbing they need was extracted from the Set binders into a shared base rather than copied: `AnimatorTriggerMonoBinder` and `AnimatorTriggerBinder` own the `IRelayCommand`/`Action` handover, the parameter check and the enabled-state notification, and each leaf supplies one line.
 - **`ListView.itemsSource`, the last UI Toolkit family.** A `ListView` owns its recycling, so this binder creates and releases nothing: it hands the list a source and tells it what changed, leaving `makeItem` and `bindItem` where they belong. An observable collection is followed and the list refreshed on every change — the granular calls need an index, and a filtered list reports none. The ViewModel's collection is wrapped rather than handed over, even when it already implements `IList`, because the list view would otherwise be able to write into it; the wrapper does not copy, so the cost is one object per binding and every mutating member throws. The source is cleared when the binding is released, so a recycled panel does not show the previous ViewModel's items for a frame.
 - **Binders for the set, the queue and the stack.** Lists and dictionaries were bound and the other three types of the collections library were not, though a set of owned ids, a queue of pending requests and a stack of open screens are all things a View shows. `ObservableCollectionMonoBinder<T>` covers all three, because that is exactly what they have in common: membership that changes, and no index worth binding to. It is deliberately not built on the list binder — a list binder's hooks carry an index, and none of these three has one that survives a change, since a queue renumbers on every dequeue. `ObservableCollectionViewModelMonoBinder` creates a view per member and releases it when the member leaves, holding views by the ViewModel they show rather than by position. What the collection already holds is replayed when it arrives, and a change reported without detail rebuilds rather than guessing at what moved.
 - **The MonoBehaviour half of the dictionary domain.** The list domain had both a serializable binder and a component one; the dictionary domain had only the serializable one, so a dictionary could be shown from a View's own field and not from a component dropped next to the objects it drives. `ObservableDictionaryMonoBinder<TKey, TValue>` dispatches exactly what its serializable twin dispatches, and `ObservableDictionaryViewModelMonoBinder` creates a view per entry through a keyed factory and releases it when the entry goes away. What the dictionary already holds is replayed when it arrives, so a View built after the data still shows it. A replacement under an existing key releases the old view and creates a new one rather than re-initialising the old: a factory may hand out a different prefab for a different ViewModel.
