+++ CHANGELOG.ru.md
@@ -13,6 +13,7 @@
 
 ### Добавлено
 
+- **Сброс триггера аниматора.** Взведённый и не потреблённый триггер остаётся взведённым, и анимация срабатывает в момент, когда её состояние становится достижимым, — часто через несколько секунд, в состоянии, которое с ней никто не связывал. Отменить это в пакете было нечем. `AnimatorResetTriggerMonoBinder` и `AnimatorResetTriggerBinder` это умеют, и стоили они почти ничего: нужная им командная обвязка не скопирована из Set-биндеров, а вынесена в общую базу — `AnimatorTriggerMonoBinder` и `AnimatorTriggerBinder` держат передачу `IRelayCommand`/`Action`, проверку параметра и уведомление о доступности, а каждый лист добавляет одну строку.
 - **`ListView.itemsSource` — последнее семейство UI Toolkit.** `ListView` сам управляет переиспользованием, поэтому этот биндер ничего не создаёт и не освобождает: он отдаёт списку источник и сообщает об изменениях, оставляя `makeItem` и `bindItem` там, где им место. За наблюдаемой коллекцией биндер следует и обновляет список на каждое изменение: точечные вызовы требуют индекса, а фильтрованный список его не сообщает. Коллекция ViewModel оборачивается, а не передаётся напрямую, даже если она уже реализует `IList`, — иначе список получил бы в неё запись; обёртка не копирует, поэтому стоит один объект на привязку, а любой изменяющий член бросает исключение. При отвязке источник очищается, чтобы переиспользованная панель не показывала кадр с элементами прежней ViewModel.
 - **Биндеры множества, очереди и стека.** Списки и словари были привязаны, а остальные три типа библиотеки коллекций — нет, хотя множество купленных id, очередь запросов и стек открытых экранов — всё это то, что показывает View. `ObservableCollectionMonoBinder<T>` закрывает все три сразу, потому что общее у них ровно это: меняющийся состав и отсутствие индекса, к которому имело бы смысл привязываться. Он намеренно не построен на биндере списка: хуки списочного несут индекс, а ни у одного из этих трёх нет индекса, переживающего изменение, — очередь перенумеровывается на каждом извлечении. `ObservableCollectionViewModelMonoBinder` создаёт view на участника и освобождает его, когда тот уходит, храня view по ViewModel, которую та показывает, а не по позиции. То, что коллекция уже содержит, проигрывается при связывании, а изменение, сообщённое без деталей, вызывает пересбор, а не догадки о том, что переместилось.
 - **MonoBehaviour-половина домена словарей.** У домена списков были и сериализуемый биндер, и компонентный; у словарей — только сериализуемый, поэтому словарь можно было показать из собственного поля View, но не компонентом, положенным рядом с объектами, которыми он управляет. `ObservableDictionaryMonoBinder<TKey, TValue>` рассылает ровно то же, что и его сериализуемый близнец, а `ObservableDictionaryViewModelMonoBinder` создаёт view на запись через фабрику по ключу и освобождает его, когда запись исчезает. То, что словарь уже содержит, проигрывается при связывании, так что View, собранная после данных, всё равно их показывает. Замена по существующему ключу освобождает старую view и создаёт новую, а не переинициализирует старую: фабрика может выдавать другой префаб для другой ViewModel.
